# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  disaster_alert_backend_internal_dto.CreateSOSRequest:
    properties:
      accuracy:
        description: Accuracy is the GPS accuracy in meters, if available from the
          mobile device.
        example: 8.5
        type: number
      address:
        description: Address is the readable location, landmark, street, or area name.
        example: Circle, Accra
        type: string
      emergencyType:
        description: |-
          EmergencyType describes the kind of emergency.

          Allowed values:
          medical, fire, security, flood, accident, other
        enum:
        - medical
        - fire
        - security
        - flood
        - accident
        - other
        example: medical
        type: string
      isLiveTracking:
        description: |-
          IsLiveTracking tells the backend whether the user wants location updates
          to continue while the emergency is active.
        example: true
        type: boolean
      latitude:
        description: Latitude is the user's current GPS latitude.
        example: 5.6037
        type: number
      longitude:
        description: Longitude is the user's current GPS longitude.
        example: -0.187
        type: number
      message:
        description: Message is an optional emergency message from the user.
        example: I need urgent help at my location.
        type: string
    required:
    - emergencyType
    - latitude
    - longitude
    type: object
...
  disaster_alert_backend_internal_dto.UpdateSOSStatusRequest:
    properties:
      status:
        description: |-
          Status is the new SOS status.

          Allowed values:
          active, assigned, resolved, cancelled
        enum:
        - active
        - assigned
        - resolved
        - cancelled
        example: resolved
        type: string
    required:
    - status
    type: object
...
info:
  contact: {}
  description: |-
    Disaster Alert API is a mission-control backend for disaster intelligence, community incident reports, SOS escalation, assistance coordination, emergency messaging, public alerts, chats, notifications, and administrator response operations.

    ================================
    QUICK START
    ================================

    Use this Swagger page to test the API safely.

    For mobile/user endpoints:
    1. Register or login using /auth/register or /auth/login.
    2. Copy the token from the login response.
    3. Click Authorize.
    4. Paste it like this: Bearer YOUR_JWT_TOKEN.
    5. Test endpoints such as /reports, /alerts, /sos, /assistance, /notifications, and /chats.

    For admin endpoints:
    1. Enter the Admin API Key under AdminApiKeyAuth.
    2. Create or validate a privilege code.
    3. Enter the full privilege code under PrivilegeCodeAuth.
    4. Test admin endpoints such as /admin/reports, /admin/alerts, /admin/sos, /admin/assistance, and /admin/chats.

    ================================
    AUTHENTICATION GUIDE
    ================================

    1. Mobile/User endpoints use BearerAuth.
    Required header:
    Authorization: Bearer <JWT_TOKEN>

    2. Basic admin management endpoints use AdminApiKeyAuth.
    Required header:
    Sigtrack-Admin-API-Key: <ADMIN_API_KEY>

    3. Privileged admin operation endpoints require BOTH AdminApiKeyAuth and PrivilegeCodeAuth.
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
paths:
...
  /admin/sos:
    get:
      description: |-
        Returns SOS emergency requests submitted by users.

        ADMIN USE:
        Admins use this endpoint to view urgent SOS emergencies that need response, assignment, or resolution.

        REQUIRED HEADERS:
        - Sigtrack-Admin-API-Key: Your admin API key.
        - X-Privilege-Code: A valid privilege code with `sos:read`.

        REQUIRED PERMISSION:
        sos:read
      produces:
      - application/json
      responses:
        "200":
          description: SOS requests fetched successfully.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid Admin API Key.
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Missing, revoked, expired, or unauthorized privilege code.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch SOS requests.
          schema:
            additionalProperties: true
            type: object
      security:
      - AdminApiKeyAuth: []
      - PrivilegeCodeAuth: []
      summary: List SOS requests
      tags:
      - Admin SOS
  /admin/sos/{id}:
    get:
      description: |-
        Returns details of a single SOS emergency request.

        WHEN TO USE THIS ENDPOINT:
        Use this when an admin wants to inspect one SOS request, view its emergency type, message, location, live tracking flag, and response status.

        REQUIRED HEADERS:
        - Sigtrack-Admin-API-Key: Your admin API key.
        - X-Privilege-Code: A valid privilege code with `sos:read`.

        REQUIRED PERMISSION:
        sos:read
      parameters:
      - description: SOS ID. This is the MongoDB ObjectID of the SOS request.
        example: 66e19b71c8f2a2b4d1234567
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SOS request fetched successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid SOS ID.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid Admin API Key.
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Missing, revoked, expired, or unauthorized privilege code.
          schema:
            additionalProperties: true
            type: object
        "404":
          description: SOS request not found.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error while fetching SOS request.
          schema:
            additionalProperties: true
            type: object
      security:
      - AdminApiKeyAuth: []
      - PrivilegeCodeAuth: []
      summary: Get one SOS request
      tags:
      - Admin SOS
  /admin/sos/{id}/status:
    put:
      consumes:
      - application/json
      description: |-
        Updates the status of an SOS emergency request from the admin dashboard.

        WHEN TO USE THIS ENDPOINT:
        Use this when an admin or responder takes over an SOS case, assigns it, resolves it, or cancels it.

        MOBILE APP EFFECT:
        The mobile app can stop SOS alarm/live tracking when status becomes resolved or cancelled, depending on your frontend logic.

        REQUIRED HEADERS:
        - Sigtrack-Admin-API-Key: Your admin API key.
        - X-Privilege-Code: A valid privilege code with `sos:update_status`.

        REQUIRED PERMISSION:
        sos:update_status

        ALLOWED STATUS VALUES:
        - active: SOS is still active.
        - assigned: A responder/admin has taken the case.
        - resolved: Emergency has been handled.
        - cancelled: SOS was cancelled.

        EXAMPLE REQUEST BODY:
        {
        "status": "assigned"
        }
      parameters:
      - description: SOS ID. This is the MongoDB ObjectID of the SOS request.
        example: 66e19b71c8f2a2b4d1234567
        in: path
        name: id
        required: true
        type: string
      - description: SOS status update payload. status is required.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disaster_alert_backend_internal_dto.UpdateSOSStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: SOS status updated successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body, validation error, or invalid status.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid Admin API Key.
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Missing, revoked, expired, or unauthorized privilege code.
          schema:
            additionalProperties: true
            type: object
        "404":
          description: SOS request not found.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error while updating SOS status.
          schema:
            additionalProperties: true
            type: object
      security:
      - AdminApiKeyAuth: []
      - PrivilegeCodeAuth: []
      summary: Update SOS status
      tags:
      - Admin SOS
...
  /sos:
    post:
      consumes:
      - application/json
      description: |-
        Allows an authenticated mobile user to trigger an urgent SOS emergency request.

        WHEN TO USE THIS ENDPOINT:
        Use this when a user is in immediate danger and needs urgent help.
        Examples: medical emergency, fire, security threat, flood danger, road accident, or other emergency.

        DIFFERENCE BETWEEN SOS AND ASSISTANCE:
        - SOS is for immediate emergency panic situations.
        - Assistance is for structured help requests such as food, shelter, evacuation, or medical support.

        AUTH REQUIRED:
        This endpoint requires a valid user JWT token.
        Click Authorize and paste: Bearer YOUR_JWT_TOKEN.

        REQUIRED FIELDS:
        - emergencyType: Type of emergency. Example: medical.
        - latitude: User's current latitude.
        - longitude: User's current longitude.

        OPTIONAL FIELDS:
        - message: Extra message from the user.
        - address: Human-readable location or landmark.
        - accuracy: GPS accuracy in meters.
        - isLiveTracking: true if location tracking should continue while SOS is active.

        EXAMPLE REQUEST BODY:
        {
        "emergencyType": "medical",
        "message": "I need urgent help at my location.",
        "latitude": 5.6037,
        "longitude": -0.1870,
        "address": "Circle, Accra",
        "accuracy": 8.5,
        "isLiveTracking": true
        }
      parameters:
      - description: SOS request payload. emergencyType, latitude, and longitude are
          required.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disaster_alert_backend_internal_dto.CreateSOSRequest'
      produces:
      - application/json
      responses:
        "201":
          description: SOS request triggered successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body, validation failed, or SOS could not be
            created.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error while triggering SOS request.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Trigger an emergency SOS
      tags:
      - User SOS
...
schemes:
...
securityDefinitions:
  AdminApiKeyAuth:
    in: header
    name: Sigtrack-Admin-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  PrivilegeCodeAuth:
    in: header
    name: X-Privilege-Code
    type: apiKey
swagger: "2.0"